EXTEND-8 : Add a listener service to manage extendable resource removal. Add an implementation of listener to remove extenders of removed resources.
git-svn-id: http://dev.lutece.paris.fr/svn/lutece/portal/trunk/plugins/extends/module-extend-actionbar@48511 bab10101-e421-0410-a517-8ce0973de3ef

+++ src/java/fr/paris/lutece/plugins/extend/modules/actionbar/service/extender/ActionbarResourceExtender.java
@@ -113,6 +113,9 @@ public class ActionbarResourceExtender extends AbstractResourceExtender
     @Override
     public void doDeleteResourceAddOn( ResourceExtenderDTO extender )
     {
-        _configService.remove( extender.getIdExtender( ) );
+        if ( extender.getIdExtender( ) > 0 )
+        {
+            _configService.remove( extender.getIdExtender( ) );
+        }
     }
 }
